Change expected object from the API
The API response failed because it seemed have expected a construct
that changed; the new construct is represented in this patch.

However, it would be a good idea to add some internal protection
for a case where the API returns either an error (for valid reasons
like the config page is not found) or for when the JSON.parse fails
to prase (if the config JSON itself is invalid). This should then
lead to some useful error for the user so the person setting this up
can fix it.

Change-Id: If31c7628038a1ec9a8bca73f87e797d35114b3b0